[test] Move test framework crypto functions to crypto/
