+++ test/functional/test_framework/crypto/ellswift.py
@@ -12,7 +12,7 @@ import os
 import random
 import unittest
 
-from test_framework.secp256k1 import FE, G, GE
+from test_framework.crypto.secp256k1 import FE, G, GE
 
 # Precomputed constant square root of -3 (mod p).
 MINUS_3_SQRT = FE(-3).sqrt()
