+++ test/functional/test_framework/crypto/siphash.py
