+++ test/functional/test_framework/messages.py
@@ -29,7 +29,7 @@ import struct
 import time
 import unittest
 
-from test_framework.siphash import siphash256
+from test_framework.crypto.siphash import siphash256
 from test_framework.util import assert_equal
 
 MAX_LOCATOR_SZ = 101
